requirements.txt
Django
mysqlclient
djangorestframework
djangorestframework-simplejwt
pillow
django-cors-headers